Refactoring; bugfixing;
