+++ blocks/targeted-block/targeted-block.js
@@ -4,7 +4,7 @@ import { getActiveRules } from '../../scripts/api/targeted-block/api.js';
 import conditionsMatched from './condition-matcher.js';
 import { readBlockConfig } from '../../scripts/aem.js';
 import { loadFragment } from '../fragment/fragment.js';
-import applyHashTagsForNamespace from '../../scripts/hashTags.js';
+import applyHashTags from '../../scripts/hashTags.js';
 
 const blocks = [];
 const displayedBlockTypes = [];
@@ -32,8 +32,7 @@ const updateTargetedBlocksVisibility = async () => {
       block.style.display = '';
     }
   });
-  // applyHashTagsForNamespace('nav');
-  applyHashTagsForNamespace('section');
+  applyHashTags('main');
 };
 
 export default function decorate(block) {
